PeerConnectionFactory: delay deletion of owned threads.
Since PeerConnection holds a ref to its creating PeerConnectionFactory, it's
possible for ~PeerConnectionFactory() to be run on its signaling thread.
Deleting a thread from within that thread is sad times, so don't do it.

It would be nicer to avoid having PeerConnection hold a ref to the factory,
and instead require the user to keep the factory alive.  Unfortunately that
changes the contract on PeerConnection{,Factory} and it's unclear how to vet
existing callers for safety.

BUG=3100
R=juberti@webrtc.org, noahric@google.com

Review URL: https://webrtc-codereview.appspot.com/11289004

git-svn-id: http://webrtc.googlecode.com/svn/trunk@5933 4adac7df-926f-26a2-2b94-8c16560cd09d