Initialize restored_packet in nack_rtx_unittest.cc.
This is to get the DrMemory Full bots to go green, this was previously
suppressed. This fix is likely hiding a real bug that should be
investigated, but it's not a regression from before. The issue should
not be closed before we figure out why this is the case and revert this
"fix".

TBR=stefan@webrtc.org
BUG=3183

Review URL: https://webrtc-codereview.appspot.com/30369004

git-svn-id: http://webrtc.googlecode.com/svn/trunk@7169 4adac7df-926f-26a2-2b94-8c16560cd09d